+++ packages/frameworks/vue/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mirrorworld/mirage.vue
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @mirrorworld/mirage.core@1.2.0
+
 ## 0.1.3
 
 ### Patch Changes